test(favorites): Move compose previews to a dedicated package, to actually exclude them from jacoco coverage.
