+++ app/src/main/kotlin/ca/rmen/android/poetassistant/main/favorites/composables/FavoritesScreenContent.kt
@@ -35,14 +35,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import ca.rmen.android.poetassistant.ExcludeFromJacocoGeneratedReport
 import ca.rmen.android.poetassistant.R
 import ca.rmen.android.poetassistant.settings.Layout
 import ca.rmen.android.poetassistant.main.Tab
 import ca.rmen.android.poetassistant.main.favorites.ExternalAppMenuItem
-import ca.rmen.android.poetassistant.theme.AppTheme
 
 const val FAVORITES_SCREEN_CONTENT_EMPTY_TAG = "FavoritesScreen_Empty"
 const val FAVORITES_SCREEN_CONTENT_LIST_TAG = "FavoritesScreen_List"
@@ -107,40 +104,3 @@ fun FavoritesScreenContent(
     }
 }
 
-@Composable
-@ExcludeFromJacocoGeneratedReport
-@Preview(showBackground = true)
-fun FavoritesScreenContentPreview() {
-    AppTheme {
-        FavoritesScreenContent(
-            favorites = listOf("apple", "banana", "cherry"),
-            layout = Layout.CLEAN,
-            externalAppMenuItemsProducer = { emptyList() },
-            onToggleFavorite = {},
-            onCopy = {},
-            onSearchInTab = { _, _ -> },
-            onExternalAppSelected = {_, _ ->},
-            onDeleteAll = {},
-            modifier = Modifier.fillMaxSize()
-        )
-    }
-}
-
-@Composable
-@ExcludeFromJacocoGeneratedReport
-@Preview(showBackground = true)
-fun FavoritesScreenContentEmptyPreview() {
-    AppTheme {
-        FavoritesScreenContent(
-            favorites = emptyList(),
-            layout = Layout.CLEAN,
-            externalAppMenuItemsProducer = { emptyList() },
-            onToggleFavorite = {},
-            onCopy = {},
-            onSearchInTab = { _, _ -> },
-            onExternalAppSelected = {_, _ ->},
-            onDeleteAll = {},
-            modifier = Modifier.fillMaxSize()
-        )
-    }
-}
